Standalone modules refactor: engine (used by manager) not using (model) utils anymore
