+++ engine/api/src/main/scala/pl/touk/nussknacker/engine/util/metrics/RateMeter.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.util.metrics
+
+trait RateMeter {
+  def mark(): Unit
+}
