+++ engine/api/src/main/scala/pl/touk/nussknacker/engine/util/service/EspTimer.scala
@@ -1,0 +1,21 @@
+package pl.touk.nussknacker.engine.util.service
+
+import pl.touk.nussknacker.engine.util.metrics.RateMeter
+
+
+case class EspTimer(rateMeter: RateMeter, histogram: Long => Unit) {
+
+  def update(nanoTimeStart: Long): Unit = {
+    val delta = System.nanoTime() - nanoTimeStart
+    rateMeter.mark()
+    histogram.apply(delta)
+  }
+}
+
+object EspTimer {
+
+  val histogramSuffix = "histogram"
+
+  val instantRateSuffix = "instantRate"
+
+}
