+++ engine/standalone/api/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/StandaloneContextLifecycle.scala
@@ -4,9 +4,9 @@ import pl.touk.nussknacker.engine.api.JobData
 
 trait StandaloneContextLifecycle {
 
-  def open(jobData: JobData, context: StandaloneContext) : Unit
+  def open(jobData: JobData, context: StandaloneContext): Unit
 
-  def close() : Unit
+  def close(): Unit
 
 
 }
