+++ engine/standalone/api/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/StandaloneSinkFactory.scala
