+++ engine/standalone/api/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/metrics/MetricsProvider.scala
@@ -1,0 +1,12 @@
+package pl.touk.nussknacker.engine.standalone.utils.metrics
+
+import cats.data.NonEmptyList
+import pl.touk.nussknacker.engine.util.service.EspTimer
+
+trait MetricsProvider {
+
+  def espTimer(processId: String, instantTimerWindowInSeconds: Long, tags: Map[String, String], name: NonEmptyList[String]): EspTimer
+
+  def close(processId: String): Unit
+
+}
