+++ engine/standalone/app/src/test/scala/pl/touk/nussknacker/engine/standalone/http/StandaloneHttpAppSpec.scala
@@ -26,7 +26,7 @@ import pl.touk.nussknacker.engine.graph.EspProcess
 import pl.touk.nussknacker.engine.marshall.ProcessMarshaller
 import pl.touk.nussknacker.engine.spel
 import pl.touk.nussknacker.engine.standalone.api.StandaloneDeploymentData
-import pl.touk.nussknacker.engine.standalone.utils.logging.StandaloneRequestResponseLogger
+import pl.touk.nussknacker.engine.standalone.http.logging.StandaloneRequestResponseLogger
 import pl.touk.nussknacker.engine.testing.ModelJarBuilder
 
 class StandaloneHttpAppSpec extends FlatSpec with Matchers with ScalatestRouteTest with BeforeAndAfterEach with FailFastCirceSupport {
