+++ engine/standalone/engine/src/main/scala/pl/touk/nussknacker/engine/standalone/metrics/NoOpMetricsProvider.scala
@@ -1,16 +1,9 @@
-package pl.touk.nussknacker.engine.standalone.utils.metrics
+package pl.touk.nussknacker.engine.standalone.metrics
 
 import cats.data.NonEmptyList
+import pl.touk.nussknacker.engine.standalone.utils.metrics.MetricsProvider
 import pl.touk.nussknacker.engine.util.metrics.GenericInstantRateMeter
 import pl.touk.nussknacker.engine.util.service.EspTimer
-
-trait MetricsProvider {
-
-  def espTimer(processId: String, instantTimerWindowInSeconds: Long, tags: Map[String, String], name: NonEmptyList[String]): EspTimer
-
-  def close(processId: String): Unit
-
-}
 
 //mainly for tests, both unit and tests from Nussknacker UI
 object NoOpMetricsProvider extends MetricsProvider {
