+++ engine/standalone/engine/src/main/scala/pl/touk/nussknacker/engine/standalone/metrics/dropwizard/InstantRateMeter.scala
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.engine.standalone.metrics.dropwizard
+
+import io.dropwizard.metrics5.Gauge
+import pl.touk.nussknacker.engine.util.metrics.GenericInstantRateMeter
+
+class InstantRateMeter extends GenericInstantRateMeter with Gauge[Double]
