+++ engine/standalone/util/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/metrics/WithEspTimers.scala
@@ -13,7 +13,3 @@ trait WithEspTimers {
   def espTimer(tags: Map[String, String], name: NonEmptyList[String]): EspTimer = context.espTimer(instantTimerWindowInSeconds, tags, name)
 
 }
-
-
-
-
