--- engine/standalone/util/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/metrics/dropwizard/InstantRateMeter.scala
@@ -1,6 +1,0 @@
-package pl.touk.nussknacker.engine.standalone.utils.metrics.dropwizard
-
-import io.dropwizard.metrics5.Gauge
-import pl.touk.nussknacker.engine.util.metrics.GenericInstantRateMeter
-
-class InstantRateMeter extends GenericInstantRateMeter with Gauge[Double]
